Added SplashScreen
I added a quick (1 second) splash screen. It is a simple image that may
be changed. The splash screen also decides which screen to load. If it
is the first run, the splash screen will load the settings. If it is
not, the splash screen will load the main page.